Dual-Interface RAG System/requirements.txt
python-dotenv
streamlit
sentence-transformers
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-chroma
langchain-groq
langchain-unstructured
unstructured
unstructured[pdf]
fpdf